Sawah - market research
To better understand the environment in which سواح will operate, both PESTEL and SWOT analyses were conducted. The PESTEL analysis explores the political, economic, social, technological, environmental, and legal factors that may influence the application’s success in Egypt’s tourism market. Meanwhile, the SWOT analysis highlights the project’s internal strengths and weaknesses, along with the external opportunities and threats it faces. Together, these tools provide a clear view of the challenges and possibilities, ensuring that سواح is built on strong strategic foundations.